Apply OOP to volume reward calculations.
